implementation of the controller layer that meets the task requirements.

+++ src/main/java/com/innovationTeam/refactoring/entity/MovieRental.java
@@ -25,5 +25,5 @@ public class MovieRental {
     private Customer customer;
 
     @Column(name = "days")
-    private int days;
+    private Long days;
 }
